% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_utils.R
\name{abridge_mx}
\alias{abridge_mx}
\title{Abridge a single-year mx schedule to 0,1,5,... using lifetable quantities}
\usage{
abridge_mx(mx, age, sex = "t", closeout = TRUE)
}
\arguments{
\item{mx}{numeric vector of single-year mortality rates (ages 0,1,2,...)}

\item{age}{numeric vector of the same length as mx, usually 0:(n-1)}

\item{sex}{character, passed to mx_to_ax(), default "t"}

\item{closeout}{logical, passed to lifetable helpers, default TRUE}
}
\value{
numeric vector of abridged mx at ages c(0, 1, 5, 10, ...)
}
\description{
Abridge a single-year mx schedule to 0,1,5,... using lifetable quantities
}
