% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andreev.R
\name{andreev}
\alias{andreev}
\title{Andreev (1982) component decomposition of e0 difference by age}
\usage{
andreev(
  mx1,
  mx2,
  age = 0:(length(mx1) - 1),
  nx = rep(1, length(mx1)),
  sex1 = "t",
  sex2 = sex1,
  closeout = TRUE
)
}
\arguments{
\item{mx1}{numeric vector of the mortality rates (central death rates) for population 1}

\item{mx2}{numeric vector of the mortality rates (central death rates) for population 2}

\item{age}{integer vector of the lower bound of each age group (currently only single ages supported)}

\item{nx}{integer vector of age intervals, default 1.}

\item{sex1}{character either the sex for population 1: Male (\code{"m"}), Female (\code{"f"}), or Total (\code{"t"})}

\item{sex2}{character either the sex for population 2: Male (\code{"m"}), Female (\code{"f"}), or Total (\code{"t"}) assumed same as \code{sex1} unless otherwise specified.}

\item{closeout}{logical. Default \code{TRUE}. Shall we use the HMD Method Protocol to close out the \code{ax} and \code{qx} values? See details.}
}
\value{
\code{cc} numeric vector with one element per age group, and which sums to the total difference in life expectancy between population 1 and 2.

vector of age-specific contributions to e0 gap
}
\description{
Implements the age-interval component formula (telescoping form)
from Andreev (1982): eps[x1,x2) = l_x1*(e2_x1 - e1_x1) - l_x2*(e2_x2 - e1_x2),
where l_ and e_ without prime are from a chosen baseline life table.
}
\details{
setting \code{closeout} to \code{TRUE} will result in value of \code{1/mx} for the final age group, of \code{ax} and a value of 1 for the closeout of \code{qx}.
}
\examples{
a   <- .001
b   <- .07
x   <- 0:100
mx1 <- a * exp(x * b)
mx2 <- a/2 * exp(x * b)
cc <- andreev(mx1, mx2, age = x)

e01 <- mx_to_e0(mx1, age = x)
e02 <- mx_to_e0(mx2, age = x)
(delta <- e02 - e01)
sum(cc)
}
\references{
\insertRef{andreev1982}{LEdecomp}
}
