% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{available_methods}
\alias{available_methods}
\title{List available decomposition methods}
\usage{
available_methods()
}
\value{
A data frame of available decomposition methods.
}
\description{
Returns a table of all implemented methods, their function name, and category.
}
\examples{
available_methods()
}
