% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopez_ruzicka.R
\name{sen_lopez_ruzicka_instantaneous}
\alias{sen_lopez_ruzicka_instantaneous}
\title{Instantaneous sensitivity via Lopez-Ruzicka decomposition}
\usage{
sen_lopez_ruzicka_instantaneous(
  mx,
  age = (1:length(mx1)) - 1,
  nx = rep(1, length(mx1)),
  sex = "t",
  perturb = 1e-06,
  closeout = TRUE
)
}
\arguments{
\item{mx}{Numeric vector of mortality rates (central death rates).}

\item{age}{integer vector of the lower bound of each age group (currently only single ages supported)}

\item{nx}{integer vector of age intervals, default 1.}

\item{sex}{Character; "m" for male, "f" for female, or "t" for total.}

\item{perturb}{Numeric; a small constant determining the perturbation size (default: 1e-6).}

\item{closeout}{logical. Default \code{TRUE}. Shall we use the HMD Method Protocol to close out the \code{ax} and \code{qx} values? See details.}
}
\description{
Estimates the sensitivity of life expectancy to small changes in mortality rates
using the Lopez-Ruzicka decomposition. This is done by perturbing the input
mortality rates up and down by a small factor and averaging the resulting directional
sensitivities to approximate a symmetric derivative.

Specifically:
\deqn{m_x^{1} = m_x \cdot \left( \frac{1}{1 - h} \right)}
\deqn{m_x^{2} = m_x \cdot (1 - h)}
and applies \code{sen_lopez_ruzicka(mx1, mx2, ...)} and \code{sen_lopez_ruzicka(mx2, mx1, ...)},
returning their average.
}
\details{
This method gives numerically identical results to
\code{sen_arriaga_sym_instantaneous()},
\code{sen_chandrasekaran_II_instantaneous()}, and
\code{sen_chandrasekaran_III_instantaneous()}.
}
\examples{
a <- 0.001
b <- 0.07
x <- 0:100
mx <- a * exp(x * b)
s <- sen_lopez_ruzicka_instantaneous(mx, age = x)
\dontrun{
plot(x, s, type = "l")
}
}
\seealso{
\code{\link{sen_lopez_ruzicka}}, \code{\link{sen_lopez_ruzicka_instantaneous2}}, \code{\link{sen_arriaga_sym_instantaneous}}, \code{\link{sen_chandrasekaran_II_instantaneous}}
}
