% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{available_methods}
\alias{available_methods}
\title{List available decomposition methods}
\usage{
available_methods(category = NULL)
}
\arguments{
\item{category}{character. one of \code{"direct"}, \code{"direct_sen"}, \code{"opt_ok"}, or \code{"general"}}
}
\value{
A data frame of available decomposition methods.
}
\description{
Returns a table of all implemented methods, their function name, and category.
}
\examples{
available_methods()
}
