% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEdecomp.R
\name{LEdecomp}
\alias{LEdecomp}
\title{Function for applying different Life-Expectancy decomposition and sensitivity methods}
\usage{
LEdecomp(
  mx1,
  mx2,
  age = (1:length(mx1)) - 1,
  nx = rep(1, length(mx1)),
  sex1 = "t",
  sex2 = sex1,
  method = c("lifetable", "arriaga", "arriaga_sym", "sen_arriaga", "sen_arriaga_sym",
    "sen_arriaga_inst", "sen_arriaga_inst2", "sen_arriaga_sym_inst",
    "sen_arriaga_sym_inst2", "chandrasekaran_ii", "sen_chandrasekaran_ii",
    "sen_chandrasekaran_ii_inst", "sen_chandrasekaran_ii_inst2", "chandrasekaran_iii",
    "sen_chandrasekaran_iii", "sen_chandrasekaran_iii_inst",
    "sen_chandrasekaran_iii_inst2", "lopez_ruzicka", "lopez_ruzicka_sym",
    "sen_lopez_ruzicka", "sen_lopez_ruzicka_sym", "sen_lopez_ruzicka_inst",
    "sen_lopez_ruzicka_inst2", 
     "horiuchi", "stepwise", "numerical"),
  closeout = TRUE,
  opt = TRUE,
  tol = 1e-10,
  Num_Intervals = 20,
  symmetrical = TRUE,
  direction = "both",
  perturb = 1e-06,
  ...
)
}
\arguments{
\item{mx1}{numeric. age-structured mortality rates for population 1.}

\item{mx2}{numeric. age-structured mortality rates for population 2, must be same length as \code{mx1}}

\item{age}{integer. lower bound of each age group, must be same length as \code{mx1}}

\item{nx}{integer vector of age intervals, default 1.}

\item{sex1}{character. \code{"m"},\code{"f"}, or \code{"t"}, affects a0 treatment.}

\item{sex2}{character. \code{"m"},\code{"f"}, or \code{"t"}, affects a0 treatment.}

\item{method}{character. \code{"lifetable"}, \code{"arriaga"}, \code{"arriaga_sym"}, \verb{"sen_arriaga"'}, \code{"sen_arriaga_sym"}, \verb{"sen_arriaga_inst"'}, \verb{"sen_arriaga_inst2"'},\code{"sen_arriaga_sym_inst"},\code{"sen_arriaga_sym_inst2"}, \code{"chandrasekaran_ii"}, \code{"sen_chandrasekaran_ii"}, \code{"sen_chandrasekaran_ii_inst"}, \code{"sen_chandrasekaran_ii_inst2"}, \code{"chandrasekaran_iii"}, \code{"sen_chandrasekaran_iii"}, \code{"sen_chandrasekaran_iii_inst"}, \code{"sen_chandrasekaran_iii_inst2"}, \code{"lopez_ruzicka"}, \code{"lopez_ruzicka_sym"}, \code{"sen_lopez_ruzicka"}, \code{"sen_lopez_ruzicka_sym"}, \code{"sen_lopez_ruzicka_inst"}, \code{"sen_lopez_ruzicka_inst2"}, \code{"sen_lopez_ruzicka_sym_inst"}, \code{"sen_lopez_ruzicka_sym_inst2"}, "\verb{horiuchi"}, \code{"stepwise"}, \code{"numerical"}}

\item{closeout}{logical. Do we handle closeout, or truncate at top age.}

\item{opt}{logical, default \code{TRUE}. For lifetable, numerical, and instantaneous sensitivity-based decomposition, shall we optimize the rate averaging to eliminate the decomposition residual?}

\item{tol}{numeric, default \code{1e-10}. Tolerance parameter for rate averaging optimization.}

\item{Num_Intervals}{numeric, default \code{FALSE}. The number of intervals to integrate over, which is used only for \code{method} = \code{"arriaga"} or the sensitvity methods of Arriaga.}

\item{symmetrical}{logical value, default \code{FALSE}. This parameter is used for \code{method} = \code{"stepwise"} and shall we want to average the results of replacing 1 with 2 and 2 with 1?}

\item{direction}{character, default \code{'up'}. This parameter is used for \code{method} = \code{"stepwise"}, and corresponds to one of \code{"up"}, \code{"down"}, or \code{"both"}.}

\item{perturb}{numeric, default \code{1e-6}, perturbation value that is a numeric constant, and a very small number.}

\item{\dots}{optional arguments passed to \code{numDeriv::grad()}}
}
\value{
A list with class \code{"LEdecomp"} including different component in the Life Expectancy decomposition method:
\itemize{
\item \code{mx1} numerical. Age-structured mortality rates for population 1.
\item \code{mx2} numerical. Age-structured mortality rates for population 2.
\item \code{age} integer. Lower bound of each age group, provided by the user. Single ages only.
\item \code{sex1} character. One of \code{"m"} (male) , \code{"f"} (female),\code{"t"} (total). Sex of first population. This affects only infant mortality.
\item \code{sex2} character. One of \code{"m"}, \code{"f"},\code{"t"}. Sex of second population. This affects only infant mortality.
\item \code{method} character. Method selected by the user to apply the life-expectancy decomposition or sensitivity analysis.
\item \code{closeout} logical. Shall we close out ax at the last age using inverse of mx (default TRUE), or assume a closed final age group?
\item \code{opt} logical. For certain sensitivity methods, shall we optimize the weighting of mortality rates at which to evaluate the sensitivity? Default TRUE
\item \code{tol} numerical. level of tolerance when optimizing sensitivity methods. Default \code{1e-10}.
\item \code{Num_Intervals} integer. For horiuchi method only. The number of intervals to discretize the method over.
\item \code{symmetrical} logical. for stepwise replacement algorithm only, do we average results of decomposing 1 vs 2 and 2 vs 1?
\item \code{direction} character. For stepwise replacement algorithm only, up age, down age, or the average of both?
\item \code{perturb} numerical. Perturbation value that is a numeric constant, and a very small number.
\item \code{sens} a vector/matrix of the sensitivity life expectancy decomposition effects that is organized in the same way as \code{mx1} and \code{mx2}.
\item \code{LE1} numerical. LE calculated from \code{mx1}.
\item \code{LE1} numerical. LE calculated from \code{mx2}.
\item \code{LEdecomp} numerical. a vector/matrix of the life expectancy decomposition effects that is organized in the same way as \code{mx1} and \code{mx2}.
}
}
\description{
A variety of exact or asymptotically exact life expectancy decomposition methods are implemented. Also, several life-expectancy descomposition sensitivity methods are implemented to answer how each age will change with an increase/decrease in life expectancy.
These include the lifetable response experiment, using three potential sensitivity functions (lifetable-based \code{ltre_lt}, an instantaneous Arriaga-derived sensitivity \code{ltre_arriaga_instantaneous}, or numerical sensitivity \code{ltre_numerical}), the directional Arriaga method \code{arriaga}, a symmetrical Arriaga method \code{arriaga_symmetrical}.
These all give similar results, except for directional \code{arriaga}, which is the most different.
}
\examples{
a <- .001
b <- .07
x <- 0:100
mx1 <- a * exp(x * b)
mx2 <- a/2 * exp(x * b)

LEdecomp(mx1,mx2,age=x,sex1='t',method = "sen_arriaga_inst")

# what methods are available?
available_methods()
}
\references{
\insertRef{arriaga1984measuring}{LEdecomp}
\insertRef{Chandrasekaran1986}{LEdecomp}
\insertRef{preston2000demography}{LEdecomp}
\insertRef{Ponnapalli2005}{LEdecomp}
}
\seealso{
\code{\link[=sen_e0_mx_lt]{sen_e0_mx_lt()}},\code{\link[=arriaga]{arriaga()}},\code{\link[=arriaga_sym]{arriaga_sym()}},
\code{\link[=sen_arriaga]{sen_arriaga()}}, \code{\link[=sen_arriaga_sym]{sen_arriaga_sym()}}
}
