% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEdecomp.R
\name{LEdecomp}
\alias{LEdecomp}
\title{Function for applying different Life-Expectancy decomposition and sensitivity methods}
\usage{
LEdecomp(
  mx1,
  mx2,
  age,
  sex1 = "t",
  sex2 = sex1,
  method = c("lifetable", "arriaga", "arriaga_sym", "sen_arriaga", "sen_arriaga_sym",
    "sen_arriaga_inst", "sen_arriaga_inst2", "chandrasekaran_ii",
    "sen_chandrasekaran_ii", "sen_chandrasekaran_ii_inst", "sen_chandrasekaran_ii_inst2",
    "chandrasekaran_iii", "sen_chandrasekaran_iii", "sen_chandrasekaran_iii_inst",
    "sen_chandrasekaran_iii_inst2", "lopez_ruzicka", "lopez_ruzicka_sym",
    "sen_lopez_ruzicka", "sen_lopez_ruzicka_sym", "sen_lopez_ruzicka_inst",
    "sen_lopez_ruzicka_inst2", "horiuchi", "stepwise", "numerical"),
  closeout = TRUE,
  opt = TRUE,
  func = FALSE,
  tol = 1e-10,
  Num_Intervals = FALSE,
  symmetrical = TRUE,
  direction = "up",
  perturb = 1e-06,
  ...
)
}
\arguments{
\item{mx1}{numeric. age-structured mortality rates for population 1.}

\item{mx2}{numeric. age-structured mortality rates for population 2, must be same length as \code{mx1}}

\item{age}{integer. lower bound of each age group, must be same length as \code{mx1}}

\item{sex1}{character. \code{"m"},\code{"f"}, or \code{"t"}, affects a0 treatment.}

\item{sex2}{character. \code{"m"},\code{"f"}, or \code{"t"}, affects a0 treatment.}

\item{method}{character. \code{"lifetable"}, \verb{"arriaga"'}, \verb{"arriaga_sym"'}, \verb{"sen_arriaga"'}, \verb{"sen_arriaga_sym"'}, \verb{"sen_arriaga_inst"'}, \verb{"sen_arriaga_inst2"'}, \verb{"chandrasekaran_ii"'}, \verb{"sen_chandrasekaran_ii"'}, \verb{"sen_chandrasekaran_ii_inst"'}, \verb{"sen_chandrasekaran_ii_inst2"'}, \verb{"chandrasekaran_iii"'}, \verb{"sen_chandrasekaran_iii"'}, \verb{"sen_chandrasekaran_iii_inst"'}, \verb{"sen_chandrasekaran_iii_inst2"'}, \verb{"lopez_ruzicka"'}, \verb{"lopez_ruzicka_sym"'}, \verb{"sen_lopez_ruzicka"'}, \verb{"sen_lopez_ruzicka_sym"'}, \verb{"sen_lopez_ruzicka_inst"'}, \verb{"sen_lopez_ruzicka_inst2"'}, \verb{"horiuchi"'}, \verb{"stepwise"'}, \code{"numerical"}}

\item{closeout}{logical. Do we handle closeout, or truncate at top age.}

\item{opt}{logical, default \code{TRUE}. For sensitivity-based decomposition, shall we optimize the rate averaging to eliminate the decomposition residual?}

\item{func}{A function specified by the user. This must be able to take the vectors \code{mx1} or \code{mx2} as its argument, and to return the value of the function. This \code{func} is only applied for \code{method} = c(\code{"stepwise"}, \code{"horiuchi"}).}

\item{tol}{numeric, default \code{1e-10}. Tolerance parameter for rate averaging optimization.}

\item{Num_Intervals}{numeric, default \code{FALSE}. The number of intervals to integrate over, which is used only for \code{method} = \code{"arriaga"} or the sensitvity methods of Arriaga.}

\item{symmetrical}{logical value, default \code{FALSE}. This parameter is used for \code{method} = \code{"stepwise"} and shall we want to average the results of replacing 1 with 2 and 2 with 1?}

\item{direction}{character, default \code{'up'}. This parameter is used for \code{method} = \code{"stepwise"}, and corresponds to one of \code{"up"}, \code{"down"}, or \code{"both"}.}

\item{perturb}{numeric, default \code{1e-6}, perturbation value that is a numeric constant, and a very small number.}

\item{\dots}{optional arguments passed to \code{numDeriv::grad()}}
}
\value{
A list with class \code{"LEdecomp"} including different component in the Life Expectancy decomposition method:
\itemize{
\item \code{mx1} age-structured mortality rates for population 1, provided by the user.
\item \code{mx2} age-structured mortality rates for population 2, provided by the user.
\item \code{age} age, integer. lower bound of each age group,  provided by the user.
\item \code{sex1} gender selected for first population, provided by the user.
\item \code{sex2} gender selected for second population, provided by the user.
\item \code{method} method selected by the user to apply the life-expectancy decomposition or sensitivity analysis.
\item \code{closeout} how the user have decide to handle closeout or truncate at top age.
\item \code{opt} for sensitivity process how to optimize
\item \code{func} function selected for estimate the life-expectancy decomposition and the sensitivity value.
\item \code{tol} level of tolerance.
\item \code{Num_Intervals} the number of intervals to integrate over.
\item \code{symmetrical} shall the user wants to average the results of replacing 1 with 2 and 2 with 1.
\item \code{direction} ----
\item \code{perturb} perturbation value that is a numeric constant, and a very small number.
\item \code{sens} a vector/matrix of the sensitivity life expectancy decomposition effects that is organized in the same way as \code{mx1} and \code{mx2}.
\item \code{LEdecomp} a vector/matrix of the life expectancy decomposition effects that is organized in the same way as \code{mx1} and \code{mx2}.
}
}
\description{
A variety of exact or asymptotically exact life expectancy decomposition methods are implemented. Also, several life-expectancy descomposition sensitivity methods are implemented to answer how each age will change with an increase/decrease in life expectancy.
These include the lifetable response experiment, using three potential sensitivity functions (lifetable-based \code{ltre_lt}, an instantaneous Arriaga-derived sensitivity \code{ltre_arriaga_instantaneous}, or numerical sensitivity \code{ltre_numerical}), the directional Arriaga method \code{arriaga}, a symmetrical Arriaga method \code{arriaga_symmetrical}.
These all give similar results, except for directional \code{arriaga}, which is the most different.
}
\examples{
a <- .001
b <- .07
x <- 0:100
mx1 <- a * exp(x * b)
mx2 <- a/2 * exp(x * b)

LEdecomp(mx1,mx2,age=x,sex1='t',method = "arriaga_instantaneous")
}
\references{
\insertRef{arriaga1984measuring}{LEdecomp}
\insertREf{Chandrasekaran1986}{LEdecomp}
\insertRef{preston2000demography}{LEdecomp}
\insertREf{Ponnapalli2005}{LEdecomp}
}
\seealso{
\\code{\link{lifetable}}, \code{\link{arriaga}},
\code{\link{arriaga_sym}}, \code{\link{sen_arriaga}}, \code{\link{sen_arriaga_sym}},
\code{\link{sen_arriaga_inst}}, \code{\link{sen_arriaga_inst2}},
\code{\link{chandrasekaran_ii}}, \code{\link{sen_chandrasekaran_ii}},
\code{\link{sen_chandrasekaran_ii_inst}}, \code{\link{sen_chandrasekaran_ii_inst2}},
\code{\link{chandrasekaran_iii}}, \code{\link{sen_chandrasekaran_iii}},
\code{\link{sen_chandrasekaran_iii_inst}}, \code{\link{sen_chandrasekaran_iii_inst2}},
\code{\link{lopez_ruzicka}}, \code{\link{lopez_ruzicka_sym}},
\code{\link{sen_lopez_ruzicka}}, \code{\link{sen_lopez_ruzicka_sym}},
\code{\link{sen_lopez_ruzicka_inst}}, \code{\link{sen_lopez_ruzicka_inst2}},
\code{\link{horiuchi}}, \code{\link{stepwise}}, \code{\link{numerical}}
}
