% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.LEdecomp.R
\name{plot.LEdecomp}
\alias{plot.LEdecomp}
\title{Function to plot the results from the LEdecomp}
\usage{
\method{plot}{LEdecomp}(x, ...)
}
\arguments{
\item{x}{\code{x} developed using function \code{LEdecomp()} which are objects of the \code{LEdecomp} class.}

\item{...}{additional arguments to show in the plot appearance.}
}
\value{
plot the different age-specific decomposition values for the selected method. This function is valid for all the methods including the sensitivities ones.
}
\description{
R function to plot the age-specific life-expectancy decomposition value using one of the different method from the \code{LEdecomp()} function.
}
\examples{

\donttest{

}
}
\references{
\insertRef{arriaga1984measuring}{coddecomp}
\insertREf{Chandrasekaran1986}{coddecomp}
\insertRef{preston2000demography}{coddecomp}
\insertREf{Ponnapalli2005}{coddecomp}
}
\seealso{
\code{\link{LEdecomp}}
}
