% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_Data.R
\docType{data}
\name{US_data}
\alias{US_data}
\title{US Mortality data}
\format{
A data frame with 4242 rows and 4 columns with class  \code{"data.frame"} and  including the following columns
\itemize{
\item \code{year} integer. Years 2000 to 2020.
\item \code{sex} character. \code{"Male"} or \code{"Female"}.
\item \code{age} integer. Ages 0, 1, ..., 99, and 100.
\item \code{mx} numeric. Mortality rates for the corresponding sex, age and year
}
}
\usage{
US_data
}
\description{
All-cause mortality rate data from the Human Mortality Database lifetables. The dataset contains information on mortality rates (mx), by year (2000-2020), sex, and (truncated) ages 0 to 100.
}
\examples{
#The dataset is executed with the following information
US_data
}
\references{
\insertRef{hmd2026}{LEdecomp}
}
\keyword{datasets}
